pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-feature-store"
version = "1.0.0"
description = "Real-time feature store for voice analytics"
authors = [
    {name = "Your Name", email = "your.email@company.com"},
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "Apache-2.0"}
keywords = ["feature-store", "voice-analytics", "real-time", "mlops"]

dependencies = [
    # Streaming & Processing
    "apache-flink==1.17.1",
    "kafka-python==2.0.2",
    "confluent-kafka==2.2.0",
    
    # Storage
    "redis==5.0.1",
    "boto3==1.28.62",
    "pyarrow==13.0.0",
    
    # API & Web
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pydantic==2.5.0",
    
    # Audio Processing
    "librosa==0.10.1",
    "soundfile==0.12.1",
    "numpy==1.24.3",
    "scipy==1.11.4",
    
    # Data & ML
    "pandas==2.1.3",
    "scikit-learn==1.3.2",
    
    # Monitoring
    "prometheus-client==0.19.0",
    "opentelemetry-api==1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "mypy==1.7.1",
    "pre-commit==3.5.0",
    "jupyter==1.0.0",
    "ipython==8.17.2"
]
spark = [
    "pyspark==3.5.0"
]

[project.scripts]
voice-feature-store-api = "voice_feature_store.api.server:main"
voice-feature-streaming = "voice_feature_store.streaming.flink_processor:main"